refactor(step3): step3 코드 수정한 부분 병합
